requirements.txt
pyspellchecker

# Dev
coverage.py